Revert "split tests for cli"
This reverts commit f159d5ee9cf7521b63311356a5661b9b95d62479.
