--- tests/cli/conftest.py
@@ -1,14 +1,0 @@
-from __future__ import annotations
-
-from typing import TYPE_CHECKING
-
-import pytest
-
-if TYPE_CHECKING:
-    from pathlib import Path
-
-
-@pytest.fixture
-def cd_tmp_path(tmp_path: Path, monkeypatch: pytest.MonkeyPatch):
-    monkeypatch.chdir(tmp_path)
-    return tmp_path
